chore(plasma-ui): migrate "{Preview,Main}.ts" after bump SB 7
