+++ packages/plasma-ui/.storybook/main.ts
@@ -1,0 +1,32 @@
+import { mergeConfig } from 'vite';
+import type { StorybookConfig } from '@storybook/react-vite';
+
+const config: StorybookConfig = {
+    staticDirs: ['public'],
+    stories: ['../src/**/*.stories.tsx', '../src/**/*.mdx'],
+    addons: ['@storybook/addon-essentials'],
+    framework: {
+        name: '@storybook/react-vite',
+        options: {},
+    },
+    core: {
+        disableTelemetry: true,
+    },
+    docs: {
+        autodocs: true,
+        defaultName: 'Docs',
+    },
+    async viteFinal(config) {
+        return mergeConfig(config, {
+            base: '',
+            resolve: {
+                dedupe: ['react', 'react-dom', 'styled-components'],
+            },
+            build: {
+                sourcemap: false,
+            },
+        });
+    },
+};
+
+export default config;
